Add dinner details popup hint, organize files
