--- src/components/add-food-components/popup/CategoryPopupElement.css
@@ -1,41 +1,0 @@
-ion-card.popupCategory {
-    width: 56px;
-    height: 56px;
-    background-size: cover;
-}
-
-ion-label.popupElementLabel > p {
-    font-size: 12px;
-}
-
-.fruit {
-    background-image: url('../../../images/fruits.jpg');
-}
-
-.veg {
-    background-image: url('../../../images/vegetables.jpg');
-}
-
-.grains {
-    background-image: url('../../../images/grains.jpg');
-}
-
-.dairy {
-    background-image: url('../../../images/dairy.jpg');
-}
-
-.protein {
-    background-image: url('../../../images/protein_food.jpg');
-}
-
-.snacks {
-    background-image: url('../../../images/snacks.jpg');
-}
-
-.coffee {
-    background-image: url('../../../images/caffeinated_drinks.jpg');
-}
-
-.softDrinks {
-    background-image: url('../../../images/soft_drinks.jpg');
-}
